add test folder and repositories

+++ tpl-service/src/application/use_cases/user/deleteUser.js
@@ -1,9 +1,18 @@
 /**
  * Delete user
+ * @async
  * @namespace application/use_cases/user/deleteUser
  * @param {string|number} id User id
  * @returns {object} User entity
  */
-export function DeleteUser(id, { userRepository }) {
+async function DeleteUser(id, { userRepository }) {
+  const exists = await userRepository.get(id);
+
+  if (!exists) {
+    throw new Error('entity not found');
+  }
+
   return userRepository.remove(id);
 }
+
+export { DeleteUser };
